+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+- 座席開放連携を追加
+
 ### Changed
 
 ### Deprecated
@@ -38,10 +40,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
-- ムビチケ券種区分(G3,G4など)を追加
+- 券種区分(G3,G4など)を追加
 
 ## v1.0.0 - 2018-10-09
 
 ### Added
 
-- ムビチケAPIをラッピングする形でファーストリリース。
+- APIをラッピングする形でファーストリリース。