+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
-- mvtk-reserve-serviceを軽量化
+- reimplement with @surfrock/soap-parser
 
 ### Deprecated